cmd(init/build): rebase
