+++ src/command/utils.rs
@@ -1,10 +1,10 @@
 //! Utility functions for commands.
 
-use std::path::{Path, PathBuf};
 use emoji;
 use error::Error;
 use progressbar::Step;
 use std::fs;
+use std::path::{Path, PathBuf};
 use PBAR;
 
 /// If an explicit path is given, then use it, otherwise assume the current
@@ -16,6 +16,15 @@ pub fn set_crate_path(path: Option<PathBuf>) -> PathBuf {
     };
 
     crate_path
+}
+
+/// Construct our `pkg` directory in the crate.
+pub fn create_pkg_dir(path: &Path, step: &Step) -> Result<(), Error> {
+    let msg = format!("{}Creating a pkg directory...", emoji::FOLDER);
+    PBAR.step(step, &msg);
+    let pkg_dir_path = path.join("pkg");
+    fs::create_dir_all(pkg_dir_path)?;
+    Ok(())
 }
 
 /// Locates the pkg directory from a specific path
